+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ### Changed
 
-- update nuvla-api dependency to v3.06
+- update nuvla-api dependency to v3.0.7
 
 ## [2.19.0] - 2022-03-08